+++ core/src/test/java/io/lionweb/serialization/FlatbuffersSerializationTest.java
@@ -218,7 +218,7 @@ public class FlatbuffersSerializationTest extends SerializationTest {
     FlatBuffersSerialization flatBuffersSerialization =
         SerializationProvider.getStandardFlatBuffersSerialization();
     SerializedChunk serializedChunk =
-        flatBuffersSerialization.serializeNodesToSerializationBlock(myInstance);
+        flatBuffersSerialization.serializeNodesToSerializationChunk(myInstance);
     assertEquals(1, serializedChunk.getClassifierInstances().size());
     SerializedClassifierInstance serializedClassifierInstance =
         serializedChunk.getClassifierInstances().get(0);
@@ -247,7 +247,7 @@ public class FlatbuffersSerializationTest extends SerializationTest {
         SerializationProvider.getStandardFlatBuffersSerialization();
     flatBuffersSerialization.enableDynamicNodes();
     SerializedChunk serializedChunk =
-        flatBuffersSerialization.serializeNodesToSerializationBlock(n1);
+        flatBuffersSerialization.serializeNodesToSerializationChunk(n1);
     assertEquals(1, serializedChunk.getLanguages().size());
     assertEquals(new UsedLanguage("l", "1"), serializedChunk.getLanguages().get(0));
 
@@ -260,7 +260,7 @@ public class FlatbuffersSerializationTest extends SerializationTest {
     assertEquals("n1", serializedA1_1.getParentNodeID());
 
     List<ClassifierInstance<?>> deserialized =
-        flatBuffersSerialization.deserializeSerializationBlock(serializedChunk);
+        flatBuffersSerialization.deserializeSerializationChunk(serializedChunk);
     assertEquals(4, deserialized.size());
     assertInstancesAreEquals(n1, deserialized.get(0));
   }
@@ -284,7 +284,7 @@ public class FlatbuffersSerializationTest extends SerializationTest {
         SerializationProvider.getStandardFlatBuffersSerialization(LionWebVersion.v2023_1);
     flatBuffersSerialization.enableDynamicNodes();
     SerializedChunk serializedChunk =
-        flatBuffersSerialization.serializeNodesToSerializationBlock(n1);
+        flatBuffersSerialization.serializeNodesToSerializationChunk(n1);
 
     assertEquals(4, serializedChunk.getClassifierInstances().size());
     SerializedClassifierInstance serializedN1 = serializedChunk.getClassifierInstances().get(0);
@@ -295,7 +295,7 @@ public class FlatbuffersSerializationTest extends SerializationTest {
     assertEquals("n1", serializedA1_1.getParentNodeID());
 
     List<ClassifierInstance<?>> deserialized =
-        flatBuffersSerialization.deserializeSerializationBlock(serializedChunk);
+        flatBuffersSerialization.deserializeSerializationChunk(serializedChunk);
     assertEquals(4, deserialized.size());
     assertInstancesAreEquals(n1, deserialized.get(0));
   }
@@ -315,7 +315,7 @@ public class FlatbuffersSerializationTest extends SerializationTest {
     FlatBuffersSerialization flatBuffersSerialization =
         SerializationProvider.getStandardFlatBuffersSerialization();
     flatBuffersSerialization.enableDynamicNodes();
-    SerializedChunk serializedChunk = flatBuffersSerialization.serializeTreeToSerializationBlock(l);
+    SerializedChunk serializedChunk = flatBuffersSerialization.serializeTreeToSerializationChunk(l);
 
     assertEquals(5, serializedChunk.getClassifierInstances().size());
     SerializedClassifierInstance serializedL = serializedChunk.getClassifierInstances().get(0);
@@ -328,7 +328,7 @@ public class FlatbuffersSerializationTest extends SerializationTest {
 
     flatBuffersSerialization.registerLanguage(metaLang);
     List<ClassifierInstance<?>> deserialized =
-        flatBuffersSerialization.deserializeSerializationBlock(serializedChunk);
+        flatBuffersSerialization.deserializeSerializationChunk(serializedChunk);
     assertEquals(5, deserialized.size());
     ClassifierInstance<?> deserializedC = deserialized.get(3);
     assertInstancesAreEquals(c, deserializedC);
@@ -360,7 +360,7 @@ public class FlatbuffersSerializationTest extends SerializationTest {
     FlatBuffersSerialization flatBuffersSerialization =
         SerializationProvider.getStandardFlatBuffersSerialization(LionWebVersion.v2023_1);
     flatBuffersSerialization.enableDynamicNodes();
-    SerializedChunk serializedChunk = flatBuffersSerialization.serializeTreeToSerializationBlock(l);
+    SerializedChunk serializedChunk = flatBuffersSerialization.serializeTreeToSerializationChunk(l);
 
     assertEquals(5, serializedChunk.getClassifierInstances().size());
     SerializedClassifierInstance serializedL = serializedChunk.getClassifierInstances().get(0);
@@ -373,7 +373,7 @@ public class FlatbuffersSerializationTest extends SerializationTest {
 
     flatBuffersSerialization.registerLanguage(metaLang);
     List<ClassifierInstance<?>> deserialized =
-        flatBuffersSerialization.deserializeSerializationBlock(serializedChunk);
+        flatBuffersSerialization.deserializeSerializationChunk(serializedChunk);
     assertEquals(5, deserialized.size());
     ClassifierInstance<?> deserializedC = deserialized.get(3);
     assertInstancesAreEquals(c, deserializedC);
@@ -405,7 +405,7 @@ public class FlatbuffersSerializationTest extends SerializationTest {
     FlatBuffersSerialization flatBuffersSerialization =
         SerializationProvider.getStandardFlatBuffersSerialization(LionWebVersion.v2023_1);
     flatBuffersSerialization.enableDynamicNodes();
-    SerializedChunk serializedChunk = flatBuffersSerialization.serializeTreeToSerializationBlock(l);
+    SerializedChunk serializedChunk = flatBuffersSerialization.serializeTreeToSerializationChunk(l);
 
     byte[] bytes = flatBuffersSerialization.serialize(serializedChunk);
     SerializedChunk deserializedChunk = flatBuffersSerialization.deserializeToChunk(bytes);
@@ -428,7 +428,7 @@ public class FlatbuffersSerializationTest extends SerializationTest {
     FlatBuffersSerialization flatBuffersSerialization =
         SerializationProvider.getStandardFlatBuffersSerialization();
     SerializedChunk serializedChunk =
-        flatBuffersSerialization.serializeNodesToSerializationBlock(n1);
+        flatBuffersSerialization.serializeNodesToSerializationChunk(n1);
 
     assertEquals(1, serializedChunk.getLanguages().size());
     assertSerializedChunkContainsLanguage(serializedChunk, l);