+++ core/src/test/java/io/lionweb/serialization/JsonSerializationTest.java
@@ -676,7 +676,7 @@ public class JsonSerializationTest extends SerializationTest {
 
     DynamicNode myInstance = new DynamicNode("instance-a", myConcept);
     JsonSerialization jsonSer = SerializationProvider.getStandardJsonSerialization();
-    SerializedChunk serializedChunk = jsonSer.serializeNodesToSerializationBlock(myInstance);
+    SerializedChunk serializedChunk = jsonSer.serializeNodesToSerializationChunk(myInstance);
     assertEquals(1, serializedChunk.getClassifierInstances().size());
     SerializedClassifierInstance serializedClassifierInstance =
         serializedChunk.getClassifierInstances().get(0);
@@ -702,7 +702,7 @@ public class JsonSerializationTest extends SerializationTest {
 
     JsonSerialization hjs = SerializationProvider.getStandardJsonSerialization();
     hjs.enableDynamicNodes();
-    SerializedChunk serializedChunk = hjs.serializeNodesToSerializationBlock(n1);
+    SerializedChunk serializedChunk = hjs.serializeNodesToSerializationChunk(n1);
 
     assertEquals(4, serializedChunk.getClassifierInstances().size());
     SerializedClassifierInstance serializedN1 = serializedChunk.getClassifierInstances().get(0);
@@ -712,7 +712,7 @@ public class JsonSerializationTest extends SerializationTest {
     SerializedClassifierInstance serializedA1_1 = serializedChunk.getClassifierInstances().get(1);
     assertEquals("n1", serializedA1_1.getParentNodeID());
 
-    List<ClassifierInstance<?>> deserialized = hjs.deserializeSerializationBlock(serializedChunk);
+    List<ClassifierInstance<?>> deserialized = hjs.deserializeSerializationChunk(serializedChunk);
     assertEquals(4, deserialized.size());
     assertInstancesAreEquals(a1_1, deserialized.get(1));
     assertEquals(deserialized.get(0), deserialized.get(1).getParent());
@@ -740,7 +740,7 @@ public class JsonSerializationTest extends SerializationTest {
 
     JsonSerialization hjs = SerializationProvider.getStandardJsonSerialization();
     hjs.enableDynamicNodes();
-    SerializedChunk serializedChunk = hjs.serializeTreeToSerializationBlock(l);
+    SerializedChunk serializedChunk = hjs.serializeTreeToSerializationChunk(l);
 
     assertEquals(5, serializedChunk.getClassifierInstances().size());
     SerializedClassifierInstance serializedL = serializedChunk.getClassifierInstances().get(0);
@@ -752,7 +752,7 @@ public class JsonSerializationTest extends SerializationTest {
     assertEquals(Arrays.asList("metaAnn_1"), serializedC.getAnnotations());
 
     hjs.registerLanguage(metaLang);
-    List<ClassifierInstance<?>> deserialized = hjs.deserializeSerializationBlock(serializedChunk);
+    List<ClassifierInstance<?>> deserialized = hjs.deserializeSerializationChunk(serializedChunk);
     assertEquals(5, deserialized.size());
     ClassifierInstance<?> deserializedC = deserialized.get(3);
     assertInstancesAreEquals(c, deserializedC);
@@ -774,7 +774,7 @@ public class JsonSerializationTest extends SerializationTest {
     ClassifierInstanceUtils.setPropertyValueByName(n1, "foo", "abc");
 
     JsonSerialization hjs = SerializationProvider.getStandardJsonSerialization();
-    SerializedChunk serializedChunk = hjs.serializeNodesToSerializationBlock(n1);
+    SerializedChunk serializedChunk = hjs.serializeNodesToSerializationChunk(n1);
 
     assertEquals(1, serializedChunk.getLanguages().size());
     assertSerializedChunkContainsLanguage(serializedChunk, l);